Remove vanity stats from README
The npms.io service went down some time this week and, in the course of
investigating sensible alternatives, it became apparent that "all time"
donwload counts for NPM packages isn't really an easy thing to retrieve.

So… rather than fight that fight for a stat that's not really
meaningful, it's much much much simpler to remove all the dynamic stuff
from this repository for now and streamline the whole affair.